my-chat-app/requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
build==1.3.0
cachetools==6.2.2
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.7
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
oracledb==3.4.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
docx2pdf==0.1.8
docx2txt==0.9
dotenv==0.9.9
durationpy==0.10
ecdsa==0.19.1
effdet==0.4.1
emoji==2.15.0
et_xmlfile==2.0.0
fastapi==0.121.3
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-vision==3.11.0
google-genai==1.51.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.6.2
kubernetes==34.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.0
langchain-openai==1.1.3
langchain-text-splitters==1.1.0
langdetect==1.0.9
langsmith==0.4.59
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
msal==1.34.0
msoffcrypto-tool==5.4.2
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.2
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.20.0
onnxruntime==1.23.2
openai==2.8.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251107
pi_heif==1.1.1
pikepdf==10.0.2
pillow==12.0.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycocotools==2.0.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pypandoc==1.16.2
pyparsing==3.2.5
pypdf==6.4.2
pypdfium2==5.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
trio==0.32.0
trio-websocket==0.12.2
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
unstructured==0.18.21
unstructured-client==0.42.4
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.2
uuid_utils==0.12.0
uvicorn==0.38.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
wrapt==2.0.1
wsproto==1.3.1
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
